% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2date.r
\name{c2date}
\alias{c2date}
\title{Hàm c2date}
\usage{
c2date(x, format = "\%d-\%b-\%Y", limit = 2030)
}
\arguments{
\item{X}{là string thí dụ như  "15-Feb-19"}
}
\description{
Hàm này để chuyển string ("15-Feb-19") thành định dạng ngày
}
\examples{
c2date ()
}
\keyword{date}
\keyword{strptime}
