% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{calculate_sum}
\alias{calculate_sum}
\title{Hàm này tính tất cả các giá trị tổng giữa x và y}
\usage{
calculate_sum(x, y, silent = FALSE)
}
\description{
Hàm này tính tất cả các giá trị tổng giữa x và y
}
