% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke.R
\name{check_estimate_prop}
\alias{check_estimate_prop}
\title{Các hàm dưới đây kiểm tra xem dữ liệu đầu vào có đủ điều kiện để tính toán hay không ?}
\usage{
check_estimate_prop(n, f)
}
\description{
Các hàm dưới đây kiểm tra xem dữ liệu đầu vào có đủ điều kiện để tính toán hay không ?
}
