% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke.R
\name{check_test_2_prop}
\alias{check_test_2_prop}
\title{Hàm kiểm tra điều kiện để áp dụng bài toán so sánh 2 tỷ lệ}
\usage{
check_test_2_prop(n1, n2, f1, f2)
}
\description{
Hàm kiểm tra điều kiện để áp dụng bài toán so sánh 2 tỷ lệ
}
