% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_check.R
\name{check_test_independent}
\alias{check_test_independent}
\title{Hàm kiểm tra điều kiện để áp dụng bài toán kiểm định tính độc lập}
\usage{
check_test_independent(matrix)
}
\description{
Hàm kiểm tra điều kiện để áp dụng bài toán kiểm định tính độc lập
}
