% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke.R
\name{check_test_n_prop}
\alias{check_test_n_prop}
\title{Hàm kiểm tra điều kiện để áp dụng bài toán so sánh nhiều tỷ lệ}
\usage{
check_test_n_prop(m_i, n_i)
}
\description{
Hàm kiểm tra điều kiện để áp dụng bài toán so sánh nhiều tỷ lệ
}
