% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_check.R
\name{check_test_prop}
\alias{check_test_prop}
\title{Hàm kiểm tra điều kiện để áp dụng bài toán kiểm định giả thiết cho tỷ lệ}
\usage{
check_test_prop(n, p0)
}
\description{
Hàm kiểm tra điều kiện để áp dụng bài toán kiểm định giả thiết cho tỷ lệ
}
