% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke.R
\name{data_simulate_regression}
\alias{data_simulate_regression}
\title{Hàm này tạo dữ liệu giả để xây dựng bài toán hồi quy tuyến tính đơn
Lưu ý: Các hệ số hồi quy tuyến tính truyền vào không phải các hệ số hồi quy tuyến tính cuối cùng
Hàm này return 2 giá trị, có thể nhận 2 giá trị này thông qua lệnh
library(gsubfn)
list\link{x, y} <- simulate_regression_data(...)}
\usage{
data_simulate_regression(n, min_x, max_x, b0, b1, sd_eps, round_digits)
}
\description{
Hàm này tạo dữ liệu giả để xây dựng bài toán hồi quy tuyến tính đơn
Lưu ý: Các hệ số hồi quy tuyến tính truyền vào không phải các hệ số hồi quy tuyến tính cuối cùng
Hàm này return 2 giá trị, có thể nhận 2 giá trị này thông qua lệnh
library(gsubfn)
list\link{x, y} <- simulate_regression_data(...)
}
