% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_simulate_test_goodness_of_fit}
\alias{data_simulate_test_goodness_of_fit}
\title{Hàm này dùng để tạo dữ liệu cho bài toán kiểm định sự phù hợp của k tỷ lệ}
\usage{
data_simulate_test_goodness_of_fit(
  expected,
  max_diff,
  min_diff,
  step = 50,
  silent = FALSE
)
}
\description{
Hàm này dùng để tạo dữ liệu cho bài toán kiểm định sự phù hợp của k tỷ lệ
}
