% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_simulate_test_independent}
\alias{data_simulate_test_independent}
\title{Dữ liệu cho bài toán kiểm định tính độc lập.}
\usage{
data_simulate_test_independent(expected_matrix, silent = FALSE)
}
\description{
Hàm này dùng để tạo dữ liệu cho bài toán kiểm định tính độc lập
}
