% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_prop}
\alias{estimate_prop}
\title{Ước lượng khoảng cho tỷ lệ.}
\usage{
estimate_prop(n, f, alpha, mode = "two.side", silent = FALSE)
}
\description{
Hàm này ước lượng khoảng cho tỷ lệ
Hàm trả về các giá trị z_alpha, bottom, top
}
