\name{f3}
\alias{f3}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Liet ke ten cac bien va nhan cua bien neu co}
\description{
Liet ke ten cac bien va nhan cua bien neu co - tuong tu nhu lenh describe (F3) cua Stata
}
\usage{
f3(data = data1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{ dataset muon liet ke ten bien va nhan }
}
\author{ Do Van Dung <dovandzung@gmail.com> }
\seealso{ Xem them \code{\link{lf}},\code{\link{lfl}} }
\examples{
data(miso)
f3(miso)
lf("san")
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
