% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke.R
\name{get_cut_vector}
\alias{get_cut_vector}
\title{Hàm này xây dựng cut_vector cho hàm data_simulate_regression}
\usage{
get_cut_vector(min, max, size)
}
\description{
Hàm này xây dựng cut_vector cho hàm data_simulate_regression
}
