% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_diff_matrix}
\alias{get_diff_matrix}
\title{Hàm này tạo 1 mà trận với các số ngẫu nhiên, và tổng các hàng và các cột đều bằng 0.
Hàm này có thể được sử dụng để tạo bảng dữ liệu trong các bài toán
\itemize{
\item Kiểm định sự phù hợp khi bình phương
\item Kiểm định k tỷ lệ
\item Kiểm định tính độc lập
(Bằng cách tính tần số lý thuyết, rồi công thêm ma trận này)
}}
\usage{
get_diff_matrix(row, column, max, min, step = 50)
}
\description{
Hàm này tạo 1 mà trận với các số ngẫu nhiên, và tổng các hàng và các cột đều bằng 0.
Hàm này có thể được sử dụng để tạo bảng dữ liệu trong các bài toán
\itemize{
\item Kiểm định sự phù hợp khi bình phương
\item Kiểm định k tỷ lệ
\item Kiểm định tính độc lập
(Bằng cách tính tần số lý thuyết, rồi công thêm ma trận này)
}
}
