% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke.R
\name{get_test_result}
\alias{get_test_result}
\title{Hàm này in ra kết quả của bài toán kiểm định giả thiết}
\usage{
get_test_result(test, c)
}
\description{
Hàm này in ra kết quả của bài toán kiểm định giả thiết
}
