\name{l4}
\alias{l4}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Tim kiem ten bien so}
\description{
Tim kiem trong danh sach ten bien so cua bo so lieu khi biet mot phan cua ten bien
}
\usage{
l4(varname, data = data1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{varname}{Mot phan cua ten bien so can tim kiem}
  \item{data}{Bo so lieu co ten bien so can tim kiem}
}
\details{

}
\value{
}
\references{  }
\author{ Do Van Dung <dovandzung@gmail.com> }
\note{ 
}
\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
\examples{
data(miso)
lf("san",miso)
l4("san",miso)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
