% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{linear_regression_predict}
\alias{linear_regression_predict}
\title{Hàm này thực hiện tính và trả về kết quả dự đoán giá trị
trong bài toán hồi quy tuyến tính đơn.}
\usage{
linear_regression_predict(x, y, value, silent = FALSE)
}
\description{
Hàm này thực hiện tính và trả về kết quả dự đoán giá trị
trong bài toán hồi quy tuyến tính đơn.
}
