% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlength.r
\name{nlength}
\alias{nlength}
\title{Hàm nlength}
\usage{
nlength(X)
}
\arguments{
\item{X}{vector}
}
\description{
Hàm này dể đếm tần suất và nhằm sắp xếp các giá trị từ lớn đến nhỏ
}
\examples{
nlength()
}
\keyword{length}
\keyword{vector}
