% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{print_huxtable}
\alias{print_huxtable}
\title{Hàm này dùng để tạo và in bảng dataframe, sử dụng huxtable.}
\usage{
print_huxtable(df)
}
\description{
Hàm này dùng để tạo và in bảng dataframe, sử dụng huxtable.
}
