% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke.R
\name{printf}
\alias{printf}
\title{File này bao gồm các hàm dùng để tính toán kết quả các bài toán mà người dùng hay gặp phải}
\usage{
printf(...)
}
\description{
File này bao gồm các hàm dùng để tính toán kết quả các bài toán mà người dùng hay gặp phải
}
