% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke.R
\name{sample_size_mean}
\alias{sample_size_mean}
\title{Hàm này xác định kích thước mẫu cho trường hợp ước lượng giá trị trung bình.}
\usage{
sample_size_mean(sigma, eps, alpha)
}
\description{
Hàm này xác định kích thước mẫu cho trường hợp ước lượng giá trị trung bình.
}
