% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke.R
\name{sample_size_prop_1}
\alias{sample_size_prop_1}
\title{Hàm này xác đỉnh kích thước mẫu cho trường hợp ước lượng tỷ lệ (công thức 1)
Hàm trả về các giá trị z_alpha, value - kết quả của phép tính}
\usage{
sample_size_prop_1(f, eps, alpha, silent = FALSE)
}
\description{
Hàm này xác đỉnh kích thước mẫu cho trường hợp ước lượng tỷ lệ (công thức 1)
Hàm trả về các giá trị z_alpha, value - kết quả của phép tính
}
