% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke.R
\name{sample_size_prop_2}
\alias{sample_size_prop_2}
\title{Hàm này xác định kích thước mẫu cho trường hợp ước lượng tỷ lệ (công thức 2)}
\usage{
sample_size_prop_2(eps, alpha)
}
\description{
Hàm này xác định kích thước mẫu cho trường hợp ước lượng tỷ lệ (công thức 2)
}
