% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\name{test_2_mean_norm}
\alias{test_2_mean_norm}
\title{Kiểm định giả thiết: So sánh 2 giá trị trung bình (phân bố chuẩn)}
\usage{
test_2_mean_norm(
  n1,
  n2,
  mean1,
  mean2,
  sigma1,
  sigma2,
  alpha,
  alternative = "neq",
  silent = FALSE
)
}
\description{
Hàm này so sánh 2 giá trị trung bình (dùng phân bố chuẩn)
Tham số alternative là 1 trong 3 giá trị: neq, less, greater tương ứng với 3 đối thiết.
Hàm trả về kết quả của test thống kê (test) và giá trị c
}
