% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke.R
\name{test_2_mean_t}
\alias{test_2_mean_t}
\title{Hàm này so sánh 2 giá trị trung bình (dùng phân bố Student)
Tham số mode là 1 trong 3 giá trị: neq, less, greater tương ứng với 3 đối thiết.
Hàm trả về kết quả của test thống kê (test), giá trị c và phương sai chung s}
\usage{
test_2_mean_t(n1, n2, mean1, mean2, s1, s2, alpha, mode = "neq")
}
\description{
Hàm này so sánh 2 giá trị trung bình (dùng phân bố Student)
Tham số mode là 1 trong 3 giá trị: neq, less, greater tương ứng với 3 đối thiết.
Hàm trả về kết quả của test thống kê (test), giá trị c và phương sai chung s
}
