% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{test_chi_squared}
\alias{test_chi_squared}
\title{Hàm này thực hiện kiểm định khi bình phương
actual, expect là các vector thể hiện tần số quan sát và tần số lý thuyết}
\usage{
test_chi_squared(actual, expected, alpha)
}
\description{
Hàm này thực hiện kiểm định khi bình phương
actual, expect là các vector thể hiện tần số quan sát và tần số lý thuyết
}
