% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{test_independent}
\alias{test_independent}
\title{Hàm này thực hiện kiểm định tính độc lập của 2 dấu hiệu A và B
matrix là một ma trận với các phần tử n_ij nằm trong bảng liên hợp các dấu hiệu
(Contingency Table)}
\usage{
test_independent(matrix, alpha)
}
\description{
Hàm này thực hiện kiểm định tính độc lập của 2 dấu hiệu A và B
matrix là một ma trận với các phần tử n_ij nằm trong bảng liên hợp các dấu hiệu
(Contingency Table)
}
