% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\name{test_k_prop}
\alias{test_k_prop}
\title{Kiểm định giả thiết: So sánh k tỷ lệ.}
\usage{
test_k_prop(m_i, n_i, alpha, silent = FALSE)
}
\description{
Hàm này thưc hiện so sánh k tỷ lệ.
m_i, n_i là các vector thể hiện số quan sát có đặc tính A
nào đó và tổng số quan sát trong các tập tổng thể.
Hàm trả về kết quả của test thống kê (test), giá trị c và các giá trị m (sum_m_i),
l (sum_l_i), N (sum_n_i)
}
