% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke.R
\name{test_n_prop}
\alias{test_n_prop}
\title{Hàm này thưc hiện so sánh n tỷ lệ.
m_i, n_i là các vector thể hiện số quan sát có đặc tính A
nào đó và tổng số quan sát trong các tập tổng thể.
Hàm trả về kết quả của test thống kê (test), giá trị c và các giá trị m (sum_m_i),
l (sum_l_i), N (sum_n_i)}
\usage{
test_n_prop(m_i, n_i, alpha, silent = FALSE)
}
\description{
Hàm này thưc hiện so sánh n tỷ lệ.
m_i, n_i là các vector thể hiện số quan sát có đặc tính A
nào đó và tổng số quan sát trong các tập tổng thể.
Hàm trả về kết quả của test thống kê (test), giá trị c và các giá trị m (sum_m_i),
l (sum_l_i), N (sum_n_i)
}
