% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\name{test_prop}
\alias{test_prop}
\title{Hàm này kiểm định giả thiết về xác suất
Tham số mode là 1 trong 3 giá trị: neq, less, greater tương ứng với 3 đối thiết.
Hàm trả về kết quả của test thống kê (test) và giá trị c}
\usage{
test_prop(n, f, p_0, alpha, mode = "neq", silent = FALSE)
}
\description{
Hàm này kiểm định giả thiết về xác suất
Tham số mode là 1 trong 3 giá trị: neq, less, greater tương ứng với 3 đối thiết.
Hàm trả về kết quả của test thống kê (test) và giá trị c
}
