% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcc.R
\name{zcc}
\alias{zcc}
\title{Hàm zcc}
\usage{
zcc(
  cases,
  data2,
  rev = c("neither", "rows", "columns", "both"),
  pctdig = 0,
  digits = NULL
)
}
\arguments{
\item{cases}{Bien so benh - Outcome}

\item{data2}{Mot ma tran gom nhieu cot cua cac loai phoi nhiem khac nhau}

\item{rev}{Dao nghich hang, cot, hay vua hang vua cot}

\item{pctdig}{So chu so thap phan cho phan tram}

\item{digits}{So chu so co nghia cho thong ke; mac dinh = getOption(digits)}

\item{label1}{Khi bien so khong co nhan thi su dung nhan mac dinh}

\item{percent}{Phan tram hang hay cot}
}
\description{
Hàm này cho phép thực hiện phân tích cho nghiên cứu case control
}
\examples{
data(ivf)
zcc(ivf$sinhnon,ivf[,c("gioi","tang.ha")])
zcc(ivf$sinhnon,ivf[,c("gioi","tang.ha")],rev="rows")
}
\keyword{case}
\keyword{control}
