% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_check.R
\name{check_estimate_prop}
\alias{check_estimate_prop}
\title{Hàm kiểm tra điều kiện để áp dụng công thức tính ước lượng khoảng cho tỷ lệ}
\usage{
check_estimate_prop(n, f)
}
\description{
Hàm kiểm tra điều kiện để áp dụng công thức tính ước lượng khoảng cho tỷ lệ
}
