% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_estimate_prop}
\alias{check_estimate_prop}
\title{File này gồm các hàm kiểm tra xem dữ liệu đầu vào có đủ điều kiện để tính toán hay không ?}
\usage{
check_estimate_prop(n, f)
}
\description{
File này gồm các hàm kiểm tra xem dữ liệu đầu vào có đủ điều kiện để tính toán hay không ?
}
