% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke.R
\name{data_simulate_continuous}
\alias{data_simulate_continuous}
\title{Hàm này tạo dữ liệu giả theo phân bố chuẩn từ min đến max
Hàm sẽ in ra bảng phân bố tần số ghép lớp, theo các giá trị cut cho trước
Lưu ý: dữ liệu chỉ cho giá trị mean và sd gần đúng và không bằng giá trị mean và sd truyền vào hàm}
\usage{
data_simulate_continuous(n, mean, sd, min, max, size)
}
\description{
Hàm này tạo dữ liệu giả theo phân bố chuẩn từ min đến max
Hàm sẽ in ra bảng phân bố tần số ghép lớp, theo các giá trị cut cho trước
Lưu ý: dữ liệu chỉ cho giá trị mean và sd gần đúng và không bằng giá trị mean và sd truyền vào hàm
}
