% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_mean_norm}
\alias{estimate_mean_norm}
\title{Ước lượng cho giá trị trung bình (phân bố chuẩn)}
\usage{
estimate_mean_norm(n, mean, sigma, alpha, mode = "two.side", silent = FALSE)
}
\description{
Hàm này dùng để tính ước lượng khoảng cho giá trị trung bình, dùng phân bố chuẩn tắc
Hàm trả về các giá trị z_alpha, bottom, top
mode có 3 chế độ: two.side
}
