% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke.R
\name{estimate_mean_norm}
\alias{estimate_mean_norm}
\title{Hàm này dùng để tính ước lượng khoảng cho giá trị trung bình, dùng phân bố chuẩn tắc
Hàm trả về các giá trị z_alpha, bottom, top}
\usage{
estimate_mean_norm(n, mean, sigma, alpha)
}
\description{
Hàm này dùng để tính ước lượng khoảng cho giá trị trung bình, dùng phân bố chuẩn tắc
Hàm trả về các giá trị z_alpha, bottom, top
}
