% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_mean_t}
\alias{estimate_mean_t}
\title{Ước lượng cho giá trị trung bình (phân bố Student)}
\usage{
estimate_mean_t(n, mean, s, alpha, mode = "two.side", silent = FALSE)
}
\description{
Hàm này dùng để tính ước lượng khoảng cho giá trị trung bình, dùng phân bố Student
#' Hàm trả về các giá trị t_alpha, bottom, top
mode có 3 chế độ: two.side
}
