% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke.R
\name{estimate_prop}
\alias{estimate_prop}
\title{Hàm này ước lượng khoảng cho tỷ lệ
Hàm trả về các giá trị z_alpha, bottom, top}
\usage{
estimate_prop(n, f, alpha, silent = FALSE)
}
\description{
Hàm này ước lượng khoảng cho tỷ lệ
Hàm trả về các giá trị z_alpha, bottom, top
}
