% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_var}
\alias{estimate_var}
\title{Hàm này ước lượng khoảng cho phương sai.
Hàm trả về các giá trị chi_sq_1, chi_sq_2,, bottom, top}
\usage{
estimate_var(n, s, alpha, mode = "two.side", silent = FALSE)
}
\description{
Hàm này ước lượng khoảng cho phương sai.
Hàm trả về các giá trị chi_sq_1, chi_sq_2,, bottom, top
}
