% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{linear_regression}
\alias{linear_regression}
\title{Bài toán hồi quy tuyến tính đơn.}
\usage{
linear_regression(x, y, silent = FALSE)
}
\description{
Hàm này thực hiện bài toán hồi quy tuyến tính đơn và
trả về các giá trị a, b (y = ax+b)
}
