% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thongke.R
\name{linear_regression_predict}
\alias{linear_regression_predict}
\title{Hàm này thực hiện việc dự đoán giá trị trong bài toán hồi quy tuyến tính đơn.}
\usage{
linear_regression_predict(x, y, value)
}
\description{
Hàm này thực hiện việc dự đoán giá trị trong bài toán hồi quy tuyến tính đơn.
}
