% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{sample_size_mean}
\alias{sample_size_mean}
\title{Xác định kích thước mấu (TH ước lượng cho trung bình)}
\usage{
sample_size_mean(sigma, eps, alpha, silent = FALSE)
}
\description{
Hàm này xác định kích thước mẫu cho trường hợp ước lượng giá trị trung bình.
Hàm trả về các giá trị z_alpha, value - kết quả của phép tính
}
