% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{sample_size_prop_2}
\alias{sample_size_prop_2}
\title{xác đỉnh kích thước mẫu (TH ước lượng tỷ lệ, công thức 2)}
\usage{
sample_size_prop_2(eps, alpha, silent = FALSE)
}
\description{
Hàm này xác định kích thước mẫu cho trường hợp ước lượng tỷ lệ (công thức 2)
Hàm trả về các giá trị z_alpha, value - kết quả của phép tính
}
