% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{test_mean_norm}
\alias{test_mean_norm}
\title{Hàm này kiểm định giả thiết về gái trị trung bình, dùng phân bố chuẩn
Tham số mode là 1 trong 3 giá trị: neq, less, greater tương ứng với 3 đối thiết.}
\usage{
test_mean_norm(n, mean, mean_0, sigma, alpha, mode = "neq")
}
\description{
Hàm này kiểm định giả thiết về gái trị trung bình, dùng phân bố chuẩn
Tham số mode là 1 trong 3 giá trị: neq, less, greater tương ứng với 3 đối thiết.
}
