% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{test_prop}
\alias{test_prop}
\title{Hàm này kiểm định giả thiết về xác suất
Tham số mode là 1 trong 3 giá trị: neq, less, greater tương ứng với 3 đối thiết.}
\usage{
test_prop(n, f, p_0, alpha, mode = "neq")
}
\description{
Hàm này kiểm định giả thiết về xác suất
Tham số mode là 1 trong 3 giá trị: neq, less, greater tương ứng với 3 đối thiết.
}
